// frontend/src/external/client/api/orval/generated/models/companyApplicationsListCompanyApplicationsParams.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * inselfy API
 * OpenAPI spec version: 1.0.0
 */

export type CompanyApplicationsListCompanyApplicationsParams = {
/**
 * ステータスで絞り込み
 */
status?: string;
/**
 * 求人IDで絞り込み
 */
job_posting_id?: string;
/**
 * キーワード検索
 */
keyword?: string;
/**
 * 応募日時 from（RFC3339）
 */
date_from?: string;
/**
 * 応募日時 to（RFC3339）
 */
date_to?: string;
/**
 * 取得件数
 */
limit?: number;
/**
 * オフセット
 */
offset?: number;
};