// frontend/src/external/client/api/orval/generated/models/modelsJobApplicationListResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * inselfy API
 * OpenAPI spec version: 1.0.0
 */
import type { ModelsJobApplicationResponse } from './modelsJobApplicationResponse';

/**
 * 応募一覧
 */
export interface ModelsJobApplicationListResponse {
  /** アイテム一覧 */
  items: ModelsJobApplicationResponse[];
  /**
     * 総件数
     * @minimum 0
     */
  total: number;
}